+++ pytest.ini
@@ -58,3 +58,4 @@ markers =
     api: marks an API test
     smoke: marks a smoke test
     ui: marks a UI test
+    bug: mark test as known bug
